Support embedding for local modules
